Started A* algorithm. Not working.
Added a node class, node comparator class AStar class and Test class.

Node class is just so it's easier for PriorityQueue.
NodeComparator is for the PriorityQueue.
Test class is a smaller tilemap to test for issues, uses some unit
testing.
AStar is all the logic to the AStar algorithm. Having issues with
finding the goal node.