Added slider, responsive grid, fixed tile mapping.
Fixed some issues with the tiles x and y coordinates. There were some
inconsistencies in the code mixing them up. It helped fixed some issues
with the A* algorithm.

Added some mouseover action events to each individual tile so that it
would replace the F, G, H values on the screen. Added the W
value(weighted A*) along with the slider so the user can change and
modify to see different results.

Added a clear path button to restart the A* algorithm, it isn't the
most efficient way to get rid of it but due to time constraints it's
done the way it is. Just repainting the whole grid to the original
tiles.

Added 'CTRL+S' as a hot key to export map from the grid window.