Added different heuristics and Uniform Cost Search
Fixed the A* algorithm to use the correct G score. For some reason the
SECOND time the algorithm gets run on a map/tileset it works much
better.

Added more heuristics for a total of 3 heuristics. User can choose it
from a dropdown menu. Also added a Uniform Cost Search option instead
of A*.

Now after you run an algorithm, the time it took to run the algorithm
along with details of the algorithm get printed to the console.