Started Sequential A Star
Started Sequential A Star. It seems to work but there's a problem with
the parents pointer. It cycles between two nodes saying they're each
others parents, but one of the nodes is one node away from the goal
node so I believe it's working.

Added a nodesExpanded value to keep track of the nodes expanded.

Made a bunch of methods private inside of the HeuristicSearch.

Fixed a bug with the generating of the coordinates. It was forcing the
coordinates to be on mountains.

Simplified the compareRanges method.