Sequential A Star working. Fixes with MORE Bugs.
We needed to reset the parent pointers in all of the single heuristic
searches. I created a resetMap method to initialize everything back to
their starting values so they wouldn't mess with later calculations.

Sequential A Star is now working. Going to test more for bugs. Extended
the GUI to handle the second weight value.

Cleaned up the test class a lot with better methods and ways to test
and benchmark everything. Need to add memory requirements.